pyproject.toml
[tool.poetry]
name = "goit_llm"
version = "0.1.0"
description = ""
authors = ["Fast Rabbit Team <54354764+Unfeir@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "goit_llm"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
fastapi = {extras = ["all"], version = "^0.100.0"}
uvicorn = {extras = ["standart"], version = "^0.22.0"}
asyncio = "^3.4.3"
sqlalchemy = "^2.0.18"
greenlet = "^2.0.2"
redis = "^4.6.0"
asyncpg = "^0.28.0"
alembic = "^1.11.1"
jinja2 = "^3.1.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cloudinary = "^1.33.0"
pydantic = {extras = ["email"], version = "^2.0.3"}
pydantic-settings = "^2.0.2"
pyjwt = "^2.8.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
companies = "^1.0.0"
psycopg2-binary = "^2.9.7"
pypdf2 = "^3.0.1"
ruff = "^0.0.290"
numpy = "^1.24.3"
h5py = "^3.9.0"
typing-extensions = "^4.8.0"
wheel = "^0.41.2"
transformers = "^4.33.2"
torch = "^2.0.1"


[tool.ruff]
extend-exclude = ["tests*"]
respect-gitignore = true
show-fixes = true
line-length = 140
select = [
          "E",  # pycodestyle errors
          "W",  # pycodestyle warnings
          "F",  # pyflakes
          "I",  # isort
          "C",  # flake8-comprehensions
          "B",  # flake8-bugbear
          "UP",  # pyupgrade
          "YTT",  # flake8-2020
          "ASYNC",  # flake8-async
          "ANN"  # flake8-annotations
          ]
ignore = [
          "C403",
          "B008",  # do not perform function calls in argument defaults
          "C901",  # too complex,
          "C408", # clashes with type anotation
          "ANN002",  # missing-type-args (unnecessary)
          "ANN003",  # missing-type-kwargs (unnecessary)
          "ANN101",  # missing-type-self (unnecessary)
          "ANN102",  # missing-type-cls (unnecessary)
          "ANN401"  # allow Any type
          ]

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]



[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
sphinx = "^7.0.1"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
setuptools = "^68.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.pytest.ini_options]
pythonpath = [".", "app"]
asyncio_mode="auto"
